em_copilot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adls-gen2-lite"
version = "0.1.0"
description = "Local Dockerized emulator for a practical subset of Azure Data Lake Storage Gen2"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "httpx>=0.27",
  "azure-storage-file-datalake>=12.14",
  "requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["src"]
